# KNN imputor fixes