@rofinn 2020-10-16
All review comments applied except from `@view`/`@views` since there wasn't a clear performance benefit during quick benchmarking. I'll merge when tests pass.